refactored network tests seperating integration tests and unit-tests(kind of), and the actors used in them. tell_preserialised added, added integration tests for tell_preserialised using a third variant of BigPinger Improved documentation in various places.
