+++ core/Cargo.toml
@@ -77,7 +77,6 @@ iovec 							= "0.1.1" # Match MIOs Version
 tempfile 		= "3"
 serde 			= {version = "1.0", features = ["derive"]}
 once_cell 		= "1.4"
-rand            = "0.7"
 
 [build-dependencies]
 rustc_version 	= "0.2"
