+++ core/src/dispatch/mod.rs
@@ -35,6 +35,7 @@ use std::{collections::VecDeque, io::ErrorKind, time::Duration};
 pub mod lookup;
 pub mod queue_manager;
 
+// Default values for network config.
 const RETRY_CONNECTIONS_INTERVAL: u64 = 5000;
 const MAX_RETRY_ATTEMPTS: u8 = 10;
 
@@ -61,6 +62,8 @@ pub struct NetworkConfig {
     buffer_config: BufferConfig,
     custom_allocator: Option<Arc<dyn ChunkAllocator>>,
     tcp_nodelay: bool,
+    max_connection_retry_attempts: u8,
+    connection_retry_interval: u64,
 }
 
 impl NetworkConfig {
@@ -73,6 +76,8 @@ impl NetworkConfig {
             buffer_config: BufferConfig::default(),
             custom_allocator: None,
             tcp_nodelay: false,
+            max_connection_retry_attempts: MAX_RETRY_ATTEMPTS,
+            connection_retry_interval: RETRY_CONNECTIONS_INTERVAL,
         }
     }
 
@@ -86,6 +91,8 @@ impl NetworkConfig {
             buffer_config,
             custom_allocator: None,
             tcp_nodelay: false,
+            max_connection_retry_attempts: MAX_RETRY_ATTEMPTS,
+            connection_retry_interval: RETRY_CONNECTIONS_INTERVAL,
         }
     }
 
@@ -103,6 +110,8 @@ impl NetworkConfig {
             buffer_config,
             custom_allocator: Some(custom_allocator),
             tcp_nodelay: false,
+            max_connection_retry_attempts: MAX_RETRY_ATTEMPTS,
+            connection_retry_interval: RETRY_CONNECTIONS_INTERVAL,
         }
     }
 
@@ -120,26 +129,58 @@ impl NetworkConfig {
         move |notify_ready| NetworkDispatcher::with_config(self.clone(), notify_ready)
     }
 
-    /// Returns a pointer to the configurations `BufferConfig`
+    /// Returns a pointer to the configurations [BufferConfig](net::buffers::BufferConfig).
     pub fn get_buffer_config(&self) -> &BufferConfig {
         &self.buffer_config
     }
 
-    /// Returns a pointer to the CustomAllocator option so that it can be cloned by the caller.
+    /// Sets the configurations [BufferConfig](net::buffers::BufferConfig) to `buffer_config`
+    pub fn set_buffer_config(&mut self, buffer_config: BufferConfig) -> () {
+        self.buffer_config = buffer_config;
+    }
+
+    /// Returns a pointer to the `CustomAllocator` option so that it can be cloned by the caller.
     pub fn get_custom_allocator(&self) -> &Option<Arc<dyn ChunkAllocator>> {
         &self.custom_allocator
     }
 
-    /// Reads the tcp_nodelay parameter of the NetworkConfig
+    /// Reads the `tcp_nodelay` parameter of the [NetworkConfig](NetworkConfig).
     pub fn get_tcp_nodelay(&self) -> bool {
         self.tcp_nodelay
     }
 
-    /// If set to true the Nagle algorithm will be turned off for all TCP Network-channels
+    /// If set to `true` the Nagle algorithm will be turned off for all TCP Network-channels.
+    ///
     /// Decreases network-latency at the cost of reduced throughput and increased congestion.
-    /// Default value is FALSE, i.e. Nagle algorithm is turned on by default.
+    ///
+    /// Default value is `false`, i.e. the Nagle algorithm is turned on by default.
     pub fn set_tcp_nodelay(&mut self, nodelay: bool) {
         self.tcp_nodelay = nodelay;
+    }
+
+    /// Configures how many attempts at re-establishing a connection will be made before giving up
+    /// and discarding the enqueued outgoing messages.
+    ///
+    /// Default value is 10 times.
+    pub fn set_max_connection_retry_attempts(&mut self, count: u8) {
+        self.max_connection_retry_attempts = count;
+    }
+
+    /// Returns the number of times the system will retry before giving up on a connection.
+    pub fn get_max_connection_retry_attempts(&self) -> u8 {
+        self.max_connection_retry_attempts
+    }
+
+    /// Configures how long to wait (in ms) between attempts at establishing a connection.
+    ///
+    /// Default value is 5000 ms.
+    pub fn set_connection_retry_interval(&mut self, milliseconds: u64) {
+        self.connection_retry_interval = milliseconds;
+    }
+
+    /// How long (in ms) the system will wait between attempts at re-establishing connection.
+    pub fn get_connection_retry_interval(&self) -> u64 {
+        self.connection_retry_interval
     }
 }
 
@@ -152,6 +193,8 @@ impl Default for NetworkConfig {
             buffer_config: BufferConfig::default(),
             custom_allocator: None,
             tcp_nodelay: false,
+            max_connection_retry_attempts: MAX_RETRY_ATTEMPTS,
+            connection_retry_interval: RETRY_CONNECTIONS_INTERVAL,
         }
     }
 }
@@ -353,7 +396,7 @@ impl NetworkDispatcher {
         let drain = self.retry_map.clone();
         self.retry_map.clear();
         for (addr, retry) in drain {
-            if retry < MAX_RETRY_ATTEMPTS {
+            if retry < self.cfg.max_connection_retry_attempts {
                 // Make sure we will re-request connection later
                 self.retry_map.insert(addr, retry + 1);
                 if let Some(bridge) = &self.net_bridge {
@@ -363,7 +406,7 @@ impl NetworkDispatcher {
                         "Dispatcher retrying connection to host {}, attempt {}/{}",
                         addr,
                         retry,
-                        MAX_RETRY_ATTEMPTS
+                        self.cfg.max_connection_retry_attempts
                     );
                     bridge.connect(Transport::TCP, addr).unwrap();
                 }
@@ -378,7 +421,7 @@ impl NetworkDispatcher {
             }
         }
         self.schedule_once(
-            Duration::from_millis(RETRY_CONNECTIONS_INTERVAL),
+            Duration::from_millis(self.cfg.connection_retry_interval),
             move |target, _id| {
                 target.schedule_retries();
                 Handled::Ok
@@ -814,7 +857,7 @@ impl Routable for NetMessage {
     }
 
     fn into_serialised(self, buf: &mut BufferEncoder) -> Result<SerialisedFrame, SerError> {
-        crate::ser_helpers::embed_msg(self, buf).map(SerialisedFrame::Chunk)
+        crate::ser_helpers::embed_msg(self, buf).map(SerialisedFrame::ChunkLease)
     }
 
     fn into_local(self) -> Result<NetMessage, SerError> {
@@ -842,10 +885,8 @@ impl Routable for (ActorPath, ActorPath, DispatchData) {
 #[cfg(test)]
 mod tests {
     use super::{super::*, *};
-    use crate::prelude::Any;
-    use bytes::{Buf, BufMut};
-    use serde::export::Formatter;
-    use std::{fmt::Debug, thread, time::Duration};
+    use crate::prelude_test::net_test_helpers::{PingerAct, PongerAct};
+    use std::{thread, time::Duration};
 
     // replace ignore with panic cfg gate when https://github.com/rust-lang/rust/pull/74754 is merged
     #[test]
@@ -991,756 +1032,9 @@ mod tests {
     }
 
     #[test]
-    fn named_registration() {
-        const ACTOR_NAME: &str = "ponger";
-
-        let mut cfg = KompactConfig::new();
-        cfg.system_components(DeadletterBox::new, NetworkConfig::default().build());
-        let system = cfg.build().expect("KompactSystem");
-        let ponger = system.create(PongerAct::new);
-        system.start(&ponger);
-
-        let _res = system.register_by_alias(&ponger, ACTOR_NAME).wait_expect(
-            Duration::from_millis(1000),
-            "Single registration with unique alias should succeed.",
-        );
-
-        let res = system
-            .register_by_alias(&ponger, ACTOR_NAME)
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Registration never completed.");
-
-        assert_eq!(
-            res,
-            Err(RegistrationError::DuplicateEntry),
-            "Duplicate alias registration should fail."
-        );
-
-        system
-            .kill_notify(ponger)
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Ponger did not die");
-        thread::sleep(Duration::from_millis(1000));
-
-        system
-            .shutdown()
-            .expect("Kompact didn't shut down properly");
-    }
-
-    #[test]
-    /// Sets up two KompactSystems with 2x Pingers and Pongers. One Ponger is registered by UUID,
-    /// the other by a custom name. One Pinger communicates with the UUID-registered Ponger,
-    /// the other with the named Ponger. Both sets are expected to exchange PING_COUNT ping-pong
-    /// messages.
-    fn remote_delivery_to_registered_actors_eager() {
-        let (system, remote) = {
-            let system = || {
-                let mut cfg = KompactConfig::new();
-                cfg.system_components(DeadletterBox::new, NetworkConfig::default().build());
-                cfg.build().expect("KompactSystem")
-            };
-            (system(), system())
-        };
-        let (ponger_unique, pouf) = remote.create_and_register(PongerAct::new_eager);
-        let (ponger_named, ponf) = remote.create_and_register(PongerAct::new_eager);
-        let poaf = remote.register_by_alias(&ponger_named, "custom_name");
-
-        let ponger_unique_path =
-            pouf.wait_expect(Duration::from_millis(1000), "Ponger failed to register!");
-        let _ = ponf.wait_expect(Duration::from_millis(1000), "Ponger failed to register!");
-        let ponger_named_path =
-            poaf.wait_expect(Duration::from_millis(1000), "Ponger failed to register!");
-
-        let (pinger_unique, piuf) =
-            system.create_and_register(move || PingerAct::new_eager(ponger_unique_path));
-        let (pinger_named, pinf) =
-            system.create_and_register(move || PingerAct::new_eager(ponger_named_path));
-
-        piuf.wait_expect(Duration::from_millis(1000), "Pinger failed to register!");
-        pinf.wait_expect(Duration::from_millis(1000), "Pinger failed to register!");
-
-        remote.start(&ponger_unique);
-        remote.start(&ponger_named);
-        system.start(&pinger_unique);
-        system.start(&pinger_named);
-
-        // TODO maybe we could do this a bit more reliable?
-        thread::sleep(Duration::from_millis(7000));
-
-        let pingfu = system.stop_notify(&pinger_unique);
-        let pingfn = system.stop_notify(&pinger_named);
-        let pongfu = remote.kill_notify(ponger_unique);
-        let pongfn = remote.kill_notify(ponger_named);
-
-        pingfu
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Pinger never stopped!");
-        pongfu
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Ponger never died!");
-        pingfn
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Pinger never stopped!");
-        pongfn
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Ponger never died!");
-        pinger_unique.on_definition(|c| {
-            println!("pinger uniq count: {}", c.count);
-            assert_eq!(c.count, PING_COUNT);
-        });
-        pinger_named.on_definition(|c| {
-            println!("pinger named count: {}", c.count);
-            assert_eq!(c.count, PING_COUNT);
-        });
-
-        system
-            .shutdown()
-            .expect("Kompact didn't shut down properly");
-    }
-
-    #[test]
-    /// Sets up two KompactSystems, one with a BigPinger and one with a BigPonger.
-    /// BigPonger will validate the BigPing messages on reception, BigPinger counts replies
-    fn remote_delivery_bigger_than_buffer_messages_lazy_tcp() {
-        let mut net_cfg = NetworkConfig::default();
-        net_cfg.buffer_config.chunk_size(128);
-        let (system, remote) = {
-            let system = || {
-                let mut cfg = KompactConfig::new();
-                cfg.system_components(DeadletterBox::new, net_cfg.clone().build());
-                cfg.build().expect("KompactSystem")
-            };
-            (system(), system())
-        };
-        let (ponger_named, ponf) = remote.create_and_register(BigPongerAct::new_lazy);
-        let poaf = remote.register_by_alias(&ponger_named, "custom_name");
-        let _ = ponf.wait_expect(Duration::from_millis(1000), "Ponger failed to register!");
-        let ponger_named_path =
-            poaf.wait_expect(Duration::from_millis(1000), "Ponger failed to register!");
-
-        let (pinger_named, pinf) =
-            system.create_and_register(move || BigPingerAct::new_lazy(ponger_named_path, 120));
-
-        pinf.wait_expect(Duration::from_millis(1000), "Pinger failed to register!");
-
-        remote.start(&ponger_named);
-        system.start(&pinger_named);
-
-        // TODO maybe we could do this a bit more reliable?
-        thread::sleep(Duration::from_millis(15000));
-
-        let pingfn = system.stop_notify(&pinger_named);
-        let pongfn = remote.kill_notify(ponger_named);
-
-        pingfn
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Pinger never stopped!");
-        pongfn
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Ponger never died!");
-        pinger_named.on_definition(|c| {
-            assert_eq!(c.count, PING_COUNT);
-        });
-
-        system
-            .shutdown()
-            .expect("Kompact didn't shut down properly");
-    }
-
-    #[test]
-    /// Sets up two KompactSystems, one with a BigPinger and one with a BigPonger.
-    /// BigPonger will validate the BigPing messages on reception, BigPinger counts replies
-    fn remote_delivery_bigger_than_buffer_messages_eager_tcp() {
-        let mut net_cfg = NetworkConfig::default();
-        net_cfg.buffer_config.chunk_size(128);
-        let buf_cfg = &net_cfg.buffer_config;
-        let (system, remote) = {
-            let system = || {
-                let mut cfg = KompactConfig::new();
-                cfg.system_components(DeadletterBox::new, net_cfg.clone().build());
-                cfg.build().expect("KompactSystem")
-            };
-            (system(), system())
-        };
-        let (ponger_named, ponf) =
-            remote.create_and_register(|| BigPongerAct::new_eager(buf_cfg.clone()));
-        let poaf = remote.register_by_alias(&ponger_named, "custom_name");
-        let _ = ponf.wait_expect(Duration::from_millis(1000), "Ponger failed to register!");
-        let ponger_named_path =
-            poaf.wait_expect(Duration::from_millis(1000), "Ponger failed to register!");
-
-        let (pinger_named, pinf) = system.create_and_register(move || {
-            BigPingerAct::new_eager(ponger_named_path, 120, buf_cfg.clone())
-        });
-
-        pinf.wait_expect(Duration::from_millis(1000), "Pinger failed to register!");
-
-        remote.start(&ponger_named);
-        system.start(&pinger_named);
-
-        // TODO maybe we could do this a bit more reliable?
-        thread::sleep(Duration::from_millis(15000));
-
-        let pingfn = system.stop_notify(&pinger_named);
-        let pongfn = remote.kill_notify(ponger_named);
-
-        pingfn
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Pinger never stopped!");
-        pongfn
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Ponger never died!");
-        pinger_named.on_definition(|c| {
-            assert_eq!(c.count, PING_COUNT);
-        });
-
-        system
-            .shutdown()
-            .expect("Kompact didn't shut down properly");
-    }
-
-    #[test]
-    /// Sets up two KompactSystems, one with a BigPinger and one with a BigPonger.
-    /// BigPonger will validate the BigPing messages on reception, BigPinger counts replies
-    fn remote_delivery_bigger_than_buffer_messages_lazy_udp() {
-        let mut net_cfg = NetworkConfig::default();
-        net_cfg.buffer_config.chunk_size(66000);
-        let (system, remote) = {
-            let system = || {
-                let mut cfg = KompactConfig::new();
-                cfg.system_components(DeadletterBox::new, net_cfg.clone().build());
-                cfg.build().expect("KompactSystem")
-            };
-            (system(), system())
-        };
-        let (ponger_named, ponf) = remote.create_and_register(BigPongerAct::new_lazy);
-        let poaf = remote.register_by_alias(&ponger_named, "custom_name");
-        let _ = ponf.wait_expect(Duration::from_millis(1000), "Ponger failed to register!");
-        let mut ponger_named_path =
-            poaf.wait_expect(Duration::from_millis(1000), "Ponger failed to register!");
-        ponger_named_path.via_udp();
-        let (pinger_named, pinf) = system
-            .create_and_register(move || BigPingerAct::new_lazy(ponger_named_path, 1500 * 3 / 4));
-
-        pinf.wait_expect(Duration::from_millis(1000), "Pinger failed to register!");
-
-        remote.start(&ponger_named);
-        system.start(&pinger_named);
-
-        // TODO maybe we could do this a bit more reliable?
-        thread::sleep(Duration::from_millis(15000));
-
-        let pingfn = system.stop_notify(&pinger_named);
-        let pongfn = remote.kill_notify(ponger_named);
-
-        pingfn
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Pinger never stopped!");
-        pongfn
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Ponger never died!");
-        pinger_named.on_definition(|c| {
-            assert_eq!(c.count, PING_COUNT);
-        });
-
-        system
-            .shutdown()
-            .expect("Kompact didn't shut down properly");
-    }
-
-    #[test]
-    /// Sets up two KompactSystems, one with a BigPinger and one with a BigPonger.
-    /// BigPonger will validate the BigPing messages on reception, BigPinger counts replies
-    fn remote_delivery_bigger_than_buffer_messages_eager_udp() {
-        let mut net_cfg = NetworkConfig::default();
-        net_cfg.buffer_config.chunk_size(66000);
-        let buf_cfg = &net_cfg.buffer_config;
-        let (system, remote) = {
-            let system = || {
-                let mut cfg = KompactConfig::new();
-                cfg.system_components(DeadletterBox::new, net_cfg.clone().build());
-                cfg.build().expect("KompactSystem")
-            };
-            (system(), system())
-        };
-        let (ponger_named, ponf) =
-            remote.create_and_register(|| BigPongerAct::new_eager(buf_cfg.clone()));
-        let poaf = remote.register_by_alias(&ponger_named, "custom_name");
-        let _ = ponf.wait_expect(Duration::from_millis(1000), "Ponger failed to register!");
-        let mut ponger_named_path =
-            poaf.wait_expect(Duration::from_millis(1000), "Ponger failed to register!");
-        ponger_named_path.via_udp();
-        let (pinger_named, pinf) = system.create_and_register(move || {
-            BigPingerAct::new_eager(ponger_named_path, 1500 * 3 / 4, buf_cfg.clone())
-        });
-
-        pinf.wait_expect(Duration::from_millis(1000), "Pinger failed to register!");
-
-        remote.start(&ponger_named);
-        system.start(&pinger_named);
-
-        // TODO maybe we could do this a bit more reliable?
-        thread::sleep(Duration::from_millis(15000));
-
-        let pingfn = system.stop_notify(&pinger_named);
-        let pongfn = remote.kill_notify(ponger_named);
-
-        pingfn
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Pinger never stopped!");
-        pongfn
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Ponger never died!");
-        pinger_named.on_definition(|c| {
-            assert_eq!(c.count, PING_COUNT);
-        });
-
-        system
-            .shutdown()
-            .expect("Kompact didn't shut down properly");
-    }
-
-    #[test]
-    /// Sets up two KompactSystems with 2x Pingers and Pongers. One Ponger is registered by UUID,
-    /// the other by a custom name. One Pinger communicates with the UUID-registered Ponger,
-    /// the other with the named Ponger. Both sets are expected to exchange PING_COUNT ping-pong
-    /// messages.
-    fn remote_delivery_to_registered_actors_eager_mixed_udp() {
-        let (system, remote) = {
-            let system = || {
-                let mut cfg = KompactConfig::new();
-                cfg.system_components(DeadletterBox::new, NetworkConfig::default().build());
-                cfg.build().expect("KompactSystem")
-            };
-            (system(), system())
-        };
-        let (ponger_unique, pouf) = remote.create_and_register(PongerAct::new_eager);
-        let (ponger_named, ponf) = remote.create_and_register(PongerAct::new_eager);
-        let poaf = remote.register_by_alias(&ponger_named, "custom_name");
-
-        let mut ponger_unique_path =
-            pouf.wait_expect(Duration::from_millis(1000), "Ponger failed to register!");
-        ponger_unique_path.via_udp();
-        let _ = ponf.wait_expect(Duration::from_millis(1000), "Ponger failed to register!");
-        let mut ponger_named_path =
-            poaf.wait_expect(Duration::from_millis(1000), "Ponger failed to register!");
-        ponger_named_path.via_udp();
-
-        let (pinger_unique, piuf) =
-            system.create_and_register(move || PingerAct::new_eager(ponger_unique_path));
-        let (pinger_named, pinf) =
-            system.create_and_register(move || PingerAct::new_eager(ponger_named_path));
-
-        piuf.wait_expect(Duration::from_millis(1000), "Pinger failed to register!");
-        pinf.wait_expect(Duration::from_millis(1000), "Pinger failed to register!");
-
-        remote.start(&ponger_unique);
-        remote.start(&ponger_named);
-        system.start(&pinger_unique);
-        system.start(&pinger_named);
-
-        // TODO maybe we could do this a bit more reliable?
-        thread::sleep(Duration::from_millis(7000));
-
-        let pingfu = system.stop_notify(&pinger_unique);
-        let pingfn = system.stop_notify(&pinger_named);
-        let pongfu = remote.kill_notify(ponger_unique);
-        let pongfn = remote.kill_notify(ponger_named);
-
-        pingfu
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Pinger never stopped!");
-        pongfu
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Ponger never died!");
-        pingfn
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Pinger never stopped!");
-        pongfn
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Ponger never died!");
-        pinger_unique.on_definition(|c| {
-            println!("pinger uniq count: {}", c.count);
-            assert_eq!(c.count, PING_COUNT);
-        });
-        pinger_named.on_definition(|c| {
-            println!("pinger named count: {}", c.count);
-            assert_eq!(c.count, PING_COUNT);
-        });
-
-        system
-            .shutdown()
-            .expect("Kompact didn't shut down properly");
-    }
-
-    #[test]
-    /// Sets up two KompactSystems with 2x Pingers and Pongers. One Ponger is registered by UUID,
-    /// the other by a custom name. One Pinger communicates with the UUID-registered Ponger,
-    /// the other with the named Ponger. Both sets are expected to exchange PING_COUNT ping-pong
-    /// messages.
-    fn remote_delivery_to_registered_actors_lazy() {
-        let (system, remote) = {
-            let system = || {
-                let mut cfg = KompactConfig::new();
-                cfg.system_components(DeadletterBox::new, NetworkConfig::default().build());
-                cfg.build().expect("KompactSystem")
-            };
-            (system(), system())
-        };
-        let (ponger_unique, pouf) = remote.create_and_register(PongerAct::new);
-        let (ponger_named, ponf) = remote.create_and_register(PongerAct::new);
-        let poaf = remote.register_by_alias(&ponger_named, "custom_name");
-
-        pouf.wait_expect(Duration::from_millis(1000), "Ponger failed to register!");
-        ponf.wait_expect(Duration::from_millis(1000), "Ponger failed to register!");
-        poaf.wait_expect(Duration::from_millis(1000), "Ponger failed to register!");
-
-        let named_path = ActorPath::Named(NamedPath::with_system(
-            remote.system_path(),
-            vec!["custom_name".into()],
-        ));
-
-        let unique_path = ActorPath::Unique(UniquePath::with_system(
-            remote.system_path(),
-            ponger_unique.id(),
-        ));
-
-        let (pinger_unique, piuf) = system.create_and_register(move || PingerAct::new(unique_path));
-        let (pinger_named, pinf) = system.create_and_register(move || PingerAct::new(named_path));
-
-        piuf.wait_expect(Duration::from_millis(1000), "Pinger failed to register!");
-        pinf.wait_expect(Duration::from_millis(1000), "Pinger failed to register!");
-
-        remote.start(&ponger_unique);
-        remote.start(&ponger_named);
-        system.start(&pinger_unique);
-        system.start(&pinger_named);
-
-        // TODO maybe we could do this a bit more reliable?
-        thread::sleep(Duration::from_millis(7000));
-
-        let pingfu = system.stop_notify(&pinger_unique);
-        let pingfn = system.stop_notify(&pinger_named);
-        let pongfu = remote.kill_notify(ponger_unique);
-        let pongfn = remote.kill_notify(ponger_named);
-
-        pingfu
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Pinger never stopped!");
-        pongfu
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Ponger never died!");
-        pingfn
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Pinger never stopped!");
-        pongfn
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Ponger never died!");
-        pinger_unique.on_definition(|c| {
-            println!("pinger uniq count: {}", c.count);
-            assert_eq!(c.count, PING_COUNT);
-        });
-        pinger_named.on_definition(|c| {
-            println!("pinger named count: {}", c.count);
-            assert_eq!(c.count, PING_COUNT);
-        });
-
-        system
-            .shutdown()
-            .expect("Kompact didn't shut down properly");
-    }
-
-    #[test]
-    /// Sets up two KompactSystems with 2x Pingers and Pongers. One Ponger is registered by UUID,
-    /// the other by a custom name. One Pinger communicates with the UUID-registered Ponger,
-    /// the other with the named Ponger. Both sets are expected to exchange PING_COUNT ping-pong
-    /// messages.
-    fn remote_delivery_to_registered_actors_lazy_mixed_udp() {
-        let (system, remote) = {
-            let system = || {
-                let mut cfg = KompactConfig::new();
-                cfg.system_components(DeadletterBox::new, NetworkConfig::default().build());
-                cfg.build().expect("KompactSystem")
-            };
-            (system(), system())
-        };
-        let (ponger_unique, pouf) = remote.create_and_register(PongerAct::new);
-        let (ponger_named, ponf) = remote.create_and_register(PongerAct::new);
-        let poaf = remote.register_by_alias(&ponger_named, "custom_name");
-
-        let mut ponger_unique_path =
-            pouf.wait_expect(Duration::from_millis(1000), "Ponger failed to register!");
-        ponger_unique_path.via_udp();
-        let _ = ponf.wait_expect(Duration::from_millis(1000), "Ponger failed to register!");
-        let mut ponger_named_path =
-            poaf.wait_expect(Duration::from_millis(1000), "Ponger failed to register!");
-        ponger_named_path.via_udp();
-
-        let (pinger_unique, piuf) =
-            system.create_and_register(move || PingerAct::new(ponger_unique_path));
-        let (pinger_named, pinf) =
-            system.create_and_register(move || PingerAct::new(ponger_named_path));
-
-        piuf.wait_expect(Duration::from_millis(1000), "Pinger failed to register!");
-        pinf.wait_expect(Duration::from_millis(1000), "Pinger failed to register!");
-
-        remote.start(&ponger_unique);
-        remote.start(&ponger_named);
-        system.start(&pinger_unique);
-        system.start(&pinger_named);
-
-        // TODO maybe we could do this a bit more reliable?
-        thread::sleep(Duration::from_millis(7000));
-
-        let pingfu = system.stop_notify(&pinger_unique);
-        let pingfn = system.stop_notify(&pinger_named);
-        let pongfu = remote.kill_notify(ponger_unique);
-        let pongfn = remote.kill_notify(ponger_named);
-
-        pingfu
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Pinger never stopped!");
-        pongfu
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Ponger never died!");
-        pingfn
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Pinger never stopped!");
-        pongfn
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Ponger never died!");
-        pinger_unique.on_definition(|c| {
-            println!("pinger uniq count: {}", c.count);
-            assert_eq!(c.count, PING_COUNT);
-        });
-        pinger_named.on_definition(|c| {
-            println!("pinger named count: {}", c.count);
-            assert_eq!(c.count, PING_COUNT);
-        });
-
-        system
-            .shutdown()
-            .expect("Kompact didn't shut down properly");
-    }
-
-    #[test]
-    /// Sets up two KompactSystems 1 and 2a, with Named paths. It first spawns a pinger-ponger couple
-    /// The ping pong process completes, it shuts down system2a, spawns a new pinger on system1
-    /// and finally boots a new system 2b with an identical networkconfig and named actor registration
-    /// The final ping pong round should then complete as system1 automatically reconnects to system2 and
-    /// transfers the enqueued messages.
-    #[ignore]
-    fn remote_lost_and_continued_connection() {
-        let system1 = || {
-            let mut cfg = KompactConfig::new();
-            cfg.system_components(
-                DeadletterBox::new,
-                NetworkConfig::new(SocketAddr::new("127.0.0.1".parse().unwrap(), 9111)).build(),
-            );
-            cfg.build().expect("KompactSystem")
-        };
-        let system2 = || {
-            let mut cfg = KompactConfig::new();
-            cfg.system_components(
-                DeadletterBox::new,
-                NetworkConfig::new(SocketAddr::new("127.0.0.1".parse().unwrap(), 9112)).build(),
-            );
-            cfg.build().expect("KompactSystem")
-        };
-
-        // Set-up system2a
-        let system2a = system2();
-        //let (ponger_unique, pouf) = remote.create_and_register(PongerAct::new);
-        let (ponger_named, ponf) = system2a.create_and_register(PongerAct::new);
-        let poaf = system2a.register_by_alias(&ponger_named, "custom_name");
-        ponf.wait_expect(Duration::from_millis(1000), "Ponger failed to register!");
-        poaf.wait_expect(Duration::from_millis(1000), "Ponger failed to register!");
-        let named_path = ActorPath::Named(NamedPath::with_system(
-            system2a.system_path(),
-            vec!["custom_name".into()],
-        ));
-        let named_path_clone = named_path.clone();
-        // Set-up system1
-        let system1 = system1();
-        let (pinger_named, pinf) =
-            system1.create_and_register(move || PingerAct::new(named_path_clone));
-        pinf.wait_expect(Duration::from_millis(1000), "Pinger failed to register!");
-
-        system2a.start(&ponger_named);
-        system1.start(&pinger_named);
-
-        // Wait for the pingpong
-        thread::sleep(Duration::from_millis(2000));
-
-        // Assert that things are going as we expect
-        let pongfn = system2a.kill_notify(ponger_named);
-        pongfn
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Ponger never died!");
-        pinger_named.on_definition(|c| {
-            println!("pinger named count: {}", c.count);
-            assert_eq!(c.count, PING_COUNT);
-        });
-        // We now kill system2
-        system2a.shutdown().ok();
-        // Start a new pinger on system1
-        let (pinger_named2, pinf2) =
-            system1.create_and_register(move || PingerAct::new(named_path));
-        pinf2.wait_expect(Duration::from_millis(1000), "Pinger failed to register!");
-        system1.start(&pinger_named2);
-        // Wait for it to send its pings, system1 should recognize the remote address
-        thread::sleep(Duration::from_millis(1000));
-        // Assert that things are going as they should be, ping count has not increased
-        pinger_named2.on_definition(|c| {
-            println!("pinger named count: {}", c.count);
-            assert_eq!(c.count, 0);
-        });
-        // Start up system2b
-        println!("Setting up system2b");
-        let system2b = system2();
-        let (ponger_named, ponf) = system2b.create_and_register(PongerAct::new);
-        let poaf = system2b.register_by_alias(&ponger_named, "custom_name");
-        ponf.wait_expect(Duration::from_millis(1000), "Ponger failed to register!");
-        poaf.wait_expect(Duration::from_millis(1000), "Ponger failed to register!");
-        println!("Starting actor on system2b");
-        system2b.start(&ponger_named);
-
-        // We give the connection plenty of time to re-establish and transfer it's old queue
-        println!("Final sleep");
-        thread::sleep(Duration::from_millis(5000));
-        // Final assertion, did our systems re-connect without lost messages?
-        println!("Asserting");
-        pinger_named2.on_definition(|c| {
-            println!("pinger named count: {}", c.count);
-            assert_eq!(c.count, PING_COUNT);
-        });
-        println!("Shutting down");
-        system1
-            .shutdown()
-            .expect("Kompact didn't shut down properly");
-        system2b
-            .shutdown()
-            .expect("Kompact didn't shut down properly");
-    }
-
-    #[test]
-    /// Identical with `remote_lost_and_continued_connection` up to the final sleep time and assertion
-    /// system1 times out in its reconnection attempts and drops the enqueued buffers.
-    /// After indirectly asserting that the queue was dropped we start up a new pinger, and assert that it succeeds.
-    #[ignore]
-    fn remote_lost_and_dropped_connection() {
-        let system1 = || {
-            let mut cfg = KompactConfig::new();
-            cfg.system_components(
-                DeadletterBox::new,
-                NetworkConfig::new(SocketAddr::new("127.0.0.1".parse().unwrap(), 9211)).build(),
-            );
-            cfg.build().expect("KompactSystem")
-        };
-        let system2 = || {
-            let mut cfg = KompactConfig::new();
-            cfg.system_components(
-                DeadletterBox::new,
-                NetworkConfig::new(SocketAddr::new("127.0.0.1".parse().unwrap(), 9212)).build(),
-            );
-            cfg.build().expect("KompactSystem")
-        };
-
-        // Set-up system2a
-        let system2a = system2();
-        //let (ponger_unique, pouf) = remote.create_and_register(PongerAct::new);
-        let (ponger_named, ponf) = system2a.create_and_register(PongerAct::new);
-        let poaf = system2a.register_by_alias(&ponger_named, "custom_name");
-        ponf.wait_expect(Duration::from_millis(1000), "Ponger failed to register!");
-        poaf.wait_expect(Duration::from_millis(1000), "Ponger failed to register!");
-        let named_path = ActorPath::Named(NamedPath::with_system(
-            system2a.system_path(),
-            vec!["custom_name".into()],
-        ));
-        let named_path_clone = named_path.clone();
-        let named_path_clone2 = named_path.clone();
-        // Set-up system1
-        let system1 = system1();
-        let (pinger_named, pinf) =
-            system1.create_and_register(move || PingerAct::new(named_path_clone));
-        pinf.wait_expect(Duration::from_millis(1000), "Pinger failed to register!");
-
-        system2a.start(&ponger_named);
-        system1.start(&pinger_named);
-
-        // Wait for the pingpong
-        thread::sleep(Duration::from_millis(2000));
-
-        // Assert that things are going as we expect
-        let pongfn = system2a.kill_notify(ponger_named);
-        pongfn
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Ponger never died!");
-        pinger_named.on_definition(|c| {
-            println!("pinger named count: {}", c.count);
-            assert_eq!(c.count, PING_COUNT);
-        });
-        // We now kill system2
-        system2a.shutdown().ok();
-        // Start a new pinger on system1
-        let (pinger_named2, pinf2) =
-            system1.create_and_register(move || PingerAct::new(named_path));
-        pinf2.wait_expect(Duration::from_millis(1000), "Pinger failed to register!");
-        system1.start(&pinger_named2);
-        // Wait for it to send its pings, system1 should recognize the remote address
-        thread::sleep(Duration::from_millis(1000));
-        // Assert that things are going as they should be, ping count has not increased
-        pinger_named2.on_definition(|c| {
-            println!("pinger named count: {}", c.count);
-            assert_eq!(c.count, 0);
-        });
-        // Sleep long-enough that the remote connection will be dropped with its queue
-        thread::sleep(Duration::from_millis(
-            (MAX_RETRY_ATTEMPTS + 2) as u64 * RETRY_CONNECTIONS_INTERVAL,
-        ));
-        // Start up system2b
-        println!("Setting up system2b");
-        let system2b = system2();
-        let (ponger_named, ponf) = system2b.create_and_register(PongerAct::new);
-        let poaf = system2b.register_by_alias(&ponger_named, "custom_name");
-        ponf.wait_expect(Duration::from_millis(1000), "Ponger failed to register!");
-        poaf.wait_expect(Duration::from_millis(1000), "Ponger failed to register!");
-        println!("Starting actor on system2b");
-        system2b.start(&ponger_named);
-
-        // We give the connection plenty of time to re-establish and transfer it's old queue
-        thread::sleep(Duration::from_millis(5000));
-        // Final assertion, did our systems re-connect without lost messages?
-        pinger_named2.on_definition(|c| {
-            println!("pinger named count: {}", c.count);
-            assert_eq!(c.count, 0);
-        });
-
-        // This one should now succeed
-        let (pinger_named2, pinf2) =
-            system1.create_and_register(move || PingerAct::new(named_path_clone2));
-        pinf2.wait_expect(Duration::from_millis(1000), "Pinger failed to register!");
-        system1.start(&pinger_named2);
-        // Wait for it to send its pings, system1 should recognize the remote address
-        thread::sleep(Duration::from_millis(1000));
-        // Assert that things are going as they should be
-        pinger_named2.on_definition(|c| {
-            println!("pinger named count: {}", c.count);
-            assert_eq!(c.count, PING_COUNT);
-        });
-
-        system1
-            .shutdown()
-            .expect("Kompact didn't shut down properly");
-        system2b
-            .shutdown()
-            .expect("Kompact didn't shut down properly");
-    }
-
-    #[test]
-    /// Identical with `remote_lost_and_continued_connection` up to the final sleep time and assertion
-    /// system1 times out in its reconnection attempts and drops the enqueued buffers.
-    /// After indirectly asserting that the queue was dropped we start up a new pinger, and assert that it succeeds.
+    // Identical with `remote_lost_and_continued_connection` up to the final sleep time and assertion
+    // system1 times out in its reconnection attempts and drops the enqueued buffers.
+    // After indirectly asserting that the queue was dropped we start up a new pinger, and assert that it succeeds.
     fn cleanup_bufferchunks_from_dead_actors() {
         let system1 = || {
             let mut cfg = KompactConfig::new();
@@ -1762,7 +1056,7 @@ mod tests {
         // Set-up system2a
         let system2a = system2();
         //let (ponger_unique, pouf) = remote.create_and_register(PongerAct::new);
-        let (ponger_named, ponf) = system2a.create_and_register(PongerAct::new);
+        let (ponger_named, ponf) = system2a.create_and_register(PongerAct::new_lazy);
         let poaf = system2a.register_by_alias(&ponger_named, "custom_name");
         ponf.wait_expect(Duration::from_millis(1000), "Ponger failed to register!");
         poaf.wait_expect(Duration::from_millis(1000), "Ponger failed to register!");
@@ -1803,7 +1097,7 @@ mod tests {
         // Start up system2b
         println!("Setting up system2b");
         let system2b = system2();
-        let (ponger_named, ponf) = system2b.create_and_register(PongerAct::new);
+        let (ponger_named, ponf) = system2b.create_and_register(PongerAct::new_lazy);
         let poaf = system2b.register_by_alias(&ponger_named, "custom_name");
         ponf.wait_expect(Duration::from_millis(1000), "Ponger failed to register!");
         poaf.wait_expect(Duration::from_millis(1000), "Ponger failed to register!");
@@ -1829,971 +1123,5 @@ mod tests {
         system2b
             .shutdown()
             .expect("Kompact didn't shut down properly");
-    }
-
-    const PING_COUNT: u64 = 10;
-    #[test]
-    fn local_delivery() {
-        let mut cfg = KompactConfig::new();
-        cfg.system_components(DeadletterBox::new, NetworkConfig::default().build());
-        let system = cfg.build().expect("KompactSystem");
-
-        let (ponger, pof) = system.create_and_register(PongerAct::new);
-        // Construct ActorPath with system's `proto` field explicitly set to LOCAL
-        let mut ponger_path = system.actor_path_for(&ponger);
-        ponger_path.set_protocol(Transport::LOCAL);
-        let (pinger, pif) = system.create_and_register(move || PingerAct::new(ponger_path));
-
-        pof.wait_expect(Duration::from_millis(1000), "Ponger failed to register!");
-        pif.wait_expect(Duration::from_millis(1000), "Pinger failed to register!");
-
-        system.start(&ponger);
-        system.start(&pinger);
-
-        // TODO no sleeps!
-        thread::sleep(Duration::from_millis(1000));
-
-        let pingf = system.stop_notify(&pinger);
-        let pongf = system.kill_notify(ponger);
-        pingf
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Pinger never stopped!");
-        pongf
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Ponger never died!");
-        pinger.on_definition(|c| {
-            assert_eq!(c.count, PING_COUNT);
-        });
-
-        system
-            .shutdown()
-            .expect("Kompact didn't shut down properly");
-    }
-
-    #[test]
-    fn local_forwarding() {
-        let mut cfg = KompactConfig::new();
-        cfg.system_components(DeadletterBox::new, NetworkConfig::default().build());
-        let system = cfg.build().expect("KompactSystem");
-
-        let (ponger, pof) = system.create_and_register(PongerAct::new);
-        // Construct ActorPath with system's `proto` field explicitly set to LOCAL
-        let mut ponger_path =
-            pof.wait_expect(Duration::from_millis(1000), "Ponger failed to register!");
-        ponger_path.set_protocol(Transport::LOCAL);
-
-        let (forwarder, fof) = system.create_and_register(move || ForwarderAct::new(ponger_path));
-        let mut forwarder_path =
-            fof.wait_expect(Duration::from_millis(1000), "Forwarder failed to register!");
-        forwarder_path.set_protocol(Transport::LOCAL);
-
-        let (pinger, pif) = system.create_and_register(move || PingerAct::new(forwarder_path));
-        let _pinger_path =
-            pif.wait_expect(Duration::from_millis(1000), "Pinger failed to register!");
-
-        system.start(&ponger);
-        system.start(&forwarder);
-        system.start(&pinger);
-
-        // TODO no sleeps!
-        thread::sleep(Duration::from_millis(1000));
-
-        let pingf = system.kill_notify(pinger.clone()); // hold on to this ref so we can check count later
-        let forwf = system.kill_notify(forwarder);
-        let pongf = system.kill_notify(ponger);
-        pingf
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Pinger never stopped!");
-        forwf
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Forwarder never stopped!");
-        pongf
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Ponger never died!");
-        pinger.on_definition(|c| {
-            assert_eq!(c.count, PING_COUNT);
-        });
-
-        system
-            .shutdown()
-            .expect("Kompact didn't shut down properly");
-    }
-
-    #[test]
-    fn local_forwarding_eager() {
-        let mut cfg = KompactConfig::new();
-        cfg.system_components(DeadletterBox::new, NetworkConfig::default().build());
-        let system = cfg.build().expect("KompactSystem");
-
-        let (ponger, pof) = system.create_and_register(PongerAct::new);
-        // Construct ActorPath with system's `proto` field explicitly set to LOCAL
-        let mut ponger_path =
-            pof.wait_expect(Duration::from_millis(1000), "Ponger failed to register!");
-        ponger_path.set_protocol(Transport::LOCAL);
-
-        let (forwarder, fof) = system.create_and_register(move || ForwarderAct::new(ponger_path));
-        let mut forwarder_path =
-            fof.wait_expect(Duration::from_millis(1000), "Forwarder failed to register!");
-        forwarder_path.set_protocol(Transport::LOCAL);
-
-        let (pinger, pif) =
-            system.create_and_register(move || PingerAct::new_eager(forwarder_path));
-        let _pinger_path =
-            pif.wait_expect(Duration::from_millis(1000), "Pinger failed to register!");
-
-        system.start(&ponger);
-        system.start(&forwarder);
-        system.start(&pinger);
-
-        // TODO no sleeps!
-        thread::sleep(Duration::from_millis(1000));
-
-        let pingf = system.kill_notify(pinger.clone()); // hold on to this ref so we can check count later
-        let forwf = system.kill_notify(forwarder);
-        let pongf = system.kill_notify(ponger);
-        pingf
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Pinger never stopped!");
-        forwf
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Forwarder never stopped!");
-        pongf
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Ponger never died!");
-        pinger.on_definition(|c| {
-            assert_eq!(c.count, PING_COUNT);
-        });
-
-        system
-            .shutdown()
-            .expect("Kompact didn't shut down properly");
-    }
-
-    #[test]
-    fn remote_forwarding_unique() {
-        let (system1, system2, system3) = {
-            let system = || {
-                let mut cfg = KompactConfig::new();
-                cfg.system_components(DeadletterBox::new, NetworkConfig::default().build());
-                cfg.build().expect("KompactSystem")
-            };
-            (system(), system(), system())
-        };
-
-        let (ponger, pof) = system1.create_and_register(PongerAct::new);
-        let ponger_path =
-            pof.wait_expect(Duration::from_millis(1000), "Ponger failed to register!");
-
-        let (forwarder, fof) = system2.create_and_register(move || ForwarderAct::new(ponger_path));
-        let forwarder_path =
-            fof.wait_expect(Duration::from_millis(1000), "Forwarder failed to register!");
-
-        let (pinger, pif) = system3.create_and_register(move || PingerAct::new(forwarder_path));
-        let _pinger_path =
-            pif.wait_expect(Duration::from_millis(1000), "Pinger failed to register!");
-
-        system1.start(&ponger);
-        system2.start(&forwarder);
-        system3.start(&pinger);
-
-        // TODO no sleeps!
-        thread::sleep(Duration::from_millis(1000));
-
-        let pingf = system3.kill_notify(pinger.clone()); // hold on to this ref so we can check count later
-        let forwf = system2.kill_notify(forwarder);
-        let pongf = system1.kill_notify(ponger);
-        pingf
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Pinger never died!");
-        forwf
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Forwarder never died!");
-        pongf
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Ponger never died!");
-
-        pinger.on_definition(|c| {
-            assert_eq!(c.count, PING_COUNT);
-        });
-
-        system1
-            .shutdown()
-            .expect("Kompact didn't shut down properly");
-        system2
-            .shutdown()
-            .expect("Kompact didn't shut down properly");
-        system3
-            .shutdown()
-            .expect("Kompact didn't shut down properly");
-    }
-
-    #[test]
-    fn remote_forwarding_unique_two_systems() {
-        let (system1, system2) = {
-            let system = || {
-                let mut cfg = KompactConfig::new();
-                cfg.system_components(DeadletterBox::new, NetworkConfig::default().build());
-                cfg.build().expect("KompactSystem")
-            };
-            (system(), system())
-        };
-
-        let (ponger, pof) = system1.create_and_register(PongerAct::new);
-        let ponger_path =
-            pof.wait_expect(Duration::from_millis(1000), "Ponger failed to register!");
-
-        let (forwarder, fof) = system2.create_and_register(move || ForwarderAct::new(ponger_path));
-        let forwarder_path =
-            fof.wait_expect(Duration::from_millis(1000), "Forwarder failed to register!");
-
-        let (pinger, pif) = system1.create_and_register(move || PingerAct::new(forwarder_path));
-        let _pinger_path =
-            pif.wait_expect(Duration::from_millis(1000), "Pinger failed to register!");
-
-        system1.start(&ponger);
-        system2.start(&forwarder);
-        system1.start(&pinger);
-
-        // TODO no sleeps!
-        thread::sleep(Duration::from_millis(1000));
-
-        let pingf = system1.kill_notify(pinger.clone()); // hold on to this ref so we can check count later
-        let forwf = system2.kill_notify(forwarder);
-        let pongf = system1.kill_notify(ponger);
-        pingf
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Pinger never died!");
-        forwf
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Forwarder never died!");
-        pongf
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Ponger never died!");
-
-        pinger.on_definition(|c| {
-            assert_eq!(c.count, PING_COUNT);
-        });
-
-        system1
-            .shutdown()
-            .expect("Kompact didn't shut down properly");
-        system2
-            .shutdown()
-            .expect("Kompact didn't shut down properly");
-    }
-
-    #[test]
-    fn remote_forwarding_named() {
-        let (system1, system2, system3) = {
-            let system = || {
-                let mut cfg = KompactConfig::new();
-                cfg.system_components(DeadletterBox::new, NetworkConfig::default().build());
-                cfg.build().expect("KompactSystem")
-            };
-            (system(), system(), system())
-        };
-
-        let (ponger, _pof) = system1.create_and_register(PongerAct::new);
-        let pnf = system1.register_by_alias(&ponger, "ponger");
-        let ponger_path =
-            pnf.wait_expect(Duration::from_millis(1000), "Ponger failed to register!");
-
-        let (forwarder, _fof) = system2.create_and_register(move || ForwarderAct::new(ponger_path));
-        let fnf = system2.register_by_alias(&forwarder, "forwarder");
-        let forwarder_path =
-            fnf.wait_expect(Duration::from_millis(1000), "Forwarder failed to register!");
-
-        let (pinger, pif) = system3.create_and_register(move || PingerAct::new(forwarder_path));
-        let _pinger_path =
-            pif.wait_expect(Duration::from_millis(1000), "Pinger failed to register!");
-
-        system1.start(&ponger);
-        system2.start(&forwarder);
-        system3.start(&pinger);
-
-        // TODO no sleeps!
-        thread::sleep(Duration::from_millis(1000));
-
-        let pingf = system3.kill_notify(pinger.clone()); // hold on to this ref so we can check count later
-        let forwf = system2.kill_notify(forwarder);
-        let pongf = system1.kill_notify(ponger);
-        pingf
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Pinger never died!");
-        forwf
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Forwarder never died!");
-        pongf
-            .wait_timeout(Duration::from_millis(1000))
-            .expect("Ponger never died!");
-
-        pinger.on_definition(|c| {
-            assert_eq!(c.count, PING_COUNT);
-        });
-
-        system1
-            .shutdown()
-            .expect("Kompact didn't shut down properly");
-        system2
-            .shutdown()
-            .expect("Kompact didn't shut down properly");
-        system3
-            .shutdown()
-            .expect("Kompact didn't shut down properly");
-    }
-
-    #[derive(Debug, Clone)]
-    struct PingMsg {
-        i: u64,
-    }
-
-    #[derive(Debug, Clone)]
-    struct PongMsg {
-        i: u64,
-    }
-
-    struct PingPongSer;
-
-    impl PingPongSer {
-        const SID: SerId = 42;
-    }
-
-    const PING_ID: i8 = 1;
-    const PONG_ID: i8 = 2;
-
-    impl Serialiser<PingMsg> for PingPongSer {
-        fn ser_id(&self) -> SerId {
-            Self::SID
-        }
-
-        fn size_hint(&self) -> Option<usize> {
-            Some(9)
-        }
-
-        fn serialise(&self, v: &PingMsg, buf: &mut dyn BufMut) -> Result<(), SerError> {
-            buf.put_i8(PING_ID);
-            buf.put_u64(v.i);
-            Result::Ok(())
-        }
-    }
-
-    impl Serialisable for PingMsg {
-        fn ser_id(&self) -> SerId {
-            42
-        }
-
-        fn size_hint(&self) -> Option<usize> {
-            Some(9)
-        }
-
-        fn serialise(&self, buf: &mut dyn BufMut) -> Result<(), SerError> {
-            buf.put_i8(PING_ID);
-            buf.put_u64(self.i);
-            Result::Ok(())
-        }
-
-        fn local(self: Box<Self>) -> Result<Box<dyn Any + Send>, Box<dyn Serialisable>> {
-            Ok(self)
-        }
-    }
-
-    impl Serialisable for PongMsg {
-        fn ser_id(&self) -> SerId {
-            42
-        }
-
-        fn size_hint(&self) -> Option<usize> {
-            Some(9)
-        }
-
-        fn serialise(&self, buf: &mut dyn BufMut) -> Result<(), SerError> {
-            buf.put_i8(PONG_ID);
-            buf.put_u64(self.i);
-            Result::Ok(())
-        }
-
-        fn local(self: Box<Self>) -> Result<Box<dyn Any + Send>, Box<dyn Serialisable>> {
-            Ok(self)
-        }
-    }
-
-    impl Serialiser<PongMsg> for PingPongSer {
-        fn ser_id(&self) -> SerId {
-            Self::SID
-        }
-
-        fn size_hint(&self) -> Option<usize> {
-            Some(9)
-        }
-
-        fn serialise(&self, v: &PongMsg, buf: &mut dyn BufMut) -> Result<(), SerError> {
-            buf.put_i8(PONG_ID);
-            buf.put_u64(v.i);
-            Result::Ok(())
-        }
-    }
-
-    impl Deserialiser<PingMsg> for PingPongSer {
-        const SER_ID: SerId = Self::SID;
-
-        fn deserialise(buf: &mut dyn Buf) -> Result<PingMsg, SerError> {
-            if buf.remaining() < 9 {
-                return Err(SerError::InvalidData(format!(
-                    "Serialised typed has 9bytes but only {}bytes remain in buffer.",
-                    buf.remaining()
-                )));
-            }
-            match buf.get_i8() {
-                PING_ID => {
-                    let i = buf.get_u64();
-                    Ok(PingMsg { i })
-                }
-                PONG_ID => Err(SerError::InvalidType(
-                    "Found PongMsg, but expected PingMsg.".into(),
-                )),
-                id => Err(SerError::InvalidType(format!(
-                    "Found unknown id {}, but expected PingMsg.",
-                    id
-                ))),
-            }
-        }
-    }
-
-    impl Deserialiser<PongMsg> for PingPongSer {
-        const SER_ID: SerId = Self::SID;
-
-        fn deserialise(buf: &mut dyn Buf) -> Result<PongMsg, SerError> {
-            if buf.remaining() < 9 {
-                return Err(SerError::InvalidData(format!(
-                    "Serialised typed has 9bytes but only {}bytes remain in buffer.",
-                    buf.remaining()
-                )));
-            }
-            match buf.get_i8() {
-                PONG_ID => {
-                    let i = buf.get_u64();
-                    Ok(PongMsg { i })
-                }
-                PING_ID => Err(SerError::InvalidType(
-                    "Found PingMsg, but expected PongMsg.".into(),
-                )),
-                id => Err(SerError::InvalidType(format!(
-                    "Found unknown id {}, but expected PingMsg.",
-                    id
-                ))),
-            }
-        }
-    }
-
-    #[derive(ComponentDefinition)]
-    struct PingerAct {
-        ctx: ComponentContext<PingerAct>,
-        target: ActorPath,
-        count: u64,
-        eager: bool,
-    }
-
-    impl PingerAct {
-        fn new(target: ActorPath) -> PingerAct {
-            PingerAct {
-                ctx: ComponentContext::uninitialised(),
-                target,
-                count: 0,
-                eager: false,
-            }
-        }
-
-        fn new_eager(target: ActorPath) -> PingerAct {
-            PingerAct {
-                ctx: ComponentContext::uninitialised(),
-                target,
-                count: 0,
-                eager: true,
-            }
-        }
-    }
-
-    impl ComponentLifecycle for PingerAct {
-        fn on_start(&mut self) -> Handled {
-            debug!(self.ctx.log(), "Starting");
-            if self.eager {
-                self.target
-                    .tell_serialised(PingMsg { i: 0 }, self)
-                    .expect("serialise");
-            } else {
-                self.target.tell(PingMsg { i: 0 }, self);
-            }
-            Handled::Ok
-        }
-    }
-
-    impl Actor for PingerAct {
-        type Message = Never;
-
-        fn receive_local(&mut self, _pong: Self::Message) -> Handled {
-            unimplemented!();
-        }
-
-        fn receive_network(&mut self, msg: NetMessage) -> Handled {
-            match msg.try_deserialise::<PongMsg, PingPongSer>() {
-                Ok(pong) => {
-                    debug!(self.ctx.log(), "Got msg {:?}", pong);
-                    self.count += 1;
-                    if self.count < PING_COUNT {
-                        if self.eager {
-                            self.target
-                                .tell_serialised(PingMsg { i: pong.i + 1 }, self)
-                                .expect("serialise");
-                        } else {
-                            self.target.tell(PingMsg { i: pong.i + 1 }, self);
-                        }
-                    }
-                }
-                Err(e) => error!(self.ctx.log(), "Error deserialising PongMsg: {:?}", e),
-            }
-            Handled::Ok
-        }
-    }
-
-    #[derive(ComponentDefinition)]
-    struct PongerAct {
-        ctx: ComponentContext<PongerAct>,
-        eager: bool,
-    }
-
-    impl PongerAct {
-        fn new() -> PongerAct {
-            PongerAct {
-                ctx: ComponentContext::uninitialised(),
-                eager: false,
-            }
-        }
-
-        fn new_eager() -> PongerAct {
-            PongerAct {
-                ctx: ComponentContext::uninitialised(),
-                eager: true,
-            }
-        }
-    }
-
-    ignore_lifecycle!(PongerAct);
-
-    impl Actor for PongerAct {
-        type Message = Never;
-
-        fn receive_local(&mut self, _ping: Self::Message) -> Handled {
-            unimplemented!();
-        }
-
-        fn receive_network(&mut self, msg: NetMessage) -> Handled {
-            let sender = msg.sender;
-            match_deser! {msg.data; {
-                ping: PingMsg [PingPongSer] => {
-                    debug!(self.ctx.log(), "Got msg {:?} from {}", ping, sender);
-                    let pong = PongMsg { i: ping.i };
-                    if self.eager {
-                        sender
-                            .tell_serialised(pong, self)
-                            .expect("PongMsg should serialise");
-                    } else {
-                        sender.tell(pong, self);
-                    }
-                },
-                !Err(e) => error!(self.ctx.log(), "Error deserialising PingMsg: {:?}", e),
-            }}
-            Handled::Ok
-        }
-    }
-
-    #[derive(ComponentDefinition)]
-    struct ForwarderAct {
-        ctx: ComponentContext<Self>,
-        forward_to: ActorPath,
-    }
-    impl ForwarderAct {
-        fn new(forward_to: ActorPath) -> Self {
-            ForwarderAct {
-                ctx: ComponentContext::uninitialised(),
-                forward_to,
-            }
-        }
-    }
-    ignore_lifecycle!(ForwarderAct);
-    impl Actor for ForwarderAct {
-        type Message = Never;
-
-        fn receive_local(&mut self, _ping: Self::Message) -> Handled {
-            unimplemented!();
-        }
-
-        fn receive_network(&mut self, msg: NetMessage) -> Handled {
-            debug!(
-                self.ctx.log(),
-                "Forwarding some msg from {} to {}", msg.sender, self.forward_to
-            );
-            self.forward_to.forward_with_original_sender(msg, self);
-            Handled::Ok
-        }
-    }
-    // BigPing and BigPong
-    use rand::{thread_rng, Rng};
-
-    #[derive(Clone)]
-    struct BigPingMsg {
-        i: u64,
-        data: Vec<u8>,
-        sum: u64,
-    }
-
-    impl BigPingMsg {
-        pub fn new(i: u64, data_size: usize) -> Self {
-            let mut data = Vec::<u8>::with_capacity(data_size);
-            for _ in 0..data_size {
-                data.push(thread_rng().gen());
-            }
-            let mut sum: u64 = 0;
-            for d in &data {
-                sum = sum + *d as u64;
-            }
-            BigPingMsg { i, data, sum }
-        }
-
-        pub fn validate(&self) -> () {
-            let mut sum: u64 = 0;
-            for d in &self.data {
-                sum = sum + *d as u64;
-            }
-            assert_eq!(self.sum, sum, "BigPingMsg invalid, {:?}", self);
-        }
-    }
-
-    #[derive(Clone)]
-    struct BigPongMsg {
-        i: u64,
-        data: Vec<u8>,
-        sum: u64,
-    }
-
-    impl BigPongMsg {
-        pub fn new(i: u64, data_size: usize) -> BigPongMsg {
-            let mut data = Vec::<u8>::with_capacity(data_size);
-            for _ in 0..data_size {
-                data.push(thread_rng().gen());
-            }
-            let mut sum: u64 = 0;
-            for d in &data {
-                sum = sum + *d as u64;
-            }
-            BigPongMsg { i, data, sum }
-        }
-
-        pub fn validate(&self) -> () {
-            let mut sum: u64 = 0;
-            for d in &self.data {
-                sum = sum + *d as u64;
-            }
-            assert_eq!(self.sum, sum, "BigPongMsg invalid, {:?}", self);
-        }
-    }
-
-    impl Debug for BigPingMsg {
-        fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-            f.debug_struct("BigPingMsg")
-                .field("i", &self.i)
-                .field("data length", &self.data.len())
-                .field("sum", &self.sum)
-                .field("data 0", &self.data[0])
-                .field("data n", &self.data[&self.data.len() - 1])
-                .finish()
-        }
-    }
-    impl Debug for BigPongMsg {
-        fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-            f.debug_struct("BigPongMsg")
-                .field("i", &self.i)
-                .field("data length", &self.data.len())
-                .field("sum", &self.sum)
-                .field("data 0", &self.data[0])
-                .field("data n", &self.data[&self.data.len() - 1])
-                .finish()
-        }
-    }
-
-    struct BigPingPongSer;
-
-    impl BigPingPongSer {
-        const SID: SerId = 43;
-    }
-
-    impl Serialisable for BigPingMsg {
-        fn ser_id(&self) -> SerId {
-            43
-        }
-
-        fn size_hint(&self) -> Option<usize> {
-            Some(32 + self.data.len())
-        }
-
-        fn serialise(&self, buf: &mut dyn BufMut) -> Result<(), SerError> {
-            buf.put_i8(PING_ID);
-            buf.put_u64(self.i);
-            buf.put_u64(self.data.len() as u64);
-            //buf.put_slice(self.data.as_slice());
-            for d in &self.data {
-                buf.put_u8(*d);
-            }
-            buf.put_u64(self.sum);
-            Result::Ok(())
-        }
-
-        fn local(self: Box<Self>) -> Result<Box<dyn Any + Send>, Box<dyn Serialisable>> {
-            Ok(self)
-        }
-    }
-
-    impl Serialisable for BigPongMsg {
-        fn ser_id(&self) -> SerId {
-            43
-        }
-
-        fn size_hint(&self) -> Option<usize> {
-            Some(32 + self.data.len())
-        }
-
-        fn serialise(&self, buf: &mut dyn BufMut) -> Result<(), SerError> {
-            buf.put_i8(PONG_ID);
-            buf.put_u64(self.i);
-            buf.put_u64(self.data.len() as u64);
-            //buf.put_slice(self.data.as_slice());
-            for d in &self.data {
-                buf.put_u8(*d);
-            }
-            buf.put_u64(self.sum);
-            Result::Ok(())
-        }
-
-        fn local(self: Box<Self>) -> Result<Box<dyn Any + Send>, Box<dyn Serialisable>> {
-            Ok(self)
-        }
-    }
-
-    impl Deserialiser<BigPingMsg> for BigPingPongSer {
-        const SER_ID: SerId = Self::SID;
-
-        fn deserialise(buf: &mut dyn Buf) -> Result<BigPingMsg, SerError> {
-            if buf.remaining() < 18 {
-                return Err(SerError::InvalidData(format!(
-                    "Serialised typed has 18 bytes but only {} bytes remain in buffer.",
-                    buf.remaining()
-                )));
-            }
-            match buf.get_i8() {
-                PING_ID => {
-                    let i = buf.get_u64();
-                    let data_len = buf.get_u64() as usize;
-                    let mut data = Vec::with_capacity(data_len as usize);
-                    for j in 0..data_len {
-                        data.insert(j, buf.get_u8());
-                    }
-                    let sum = buf.get_u64();
-                    assert_eq!(buf.remaining(), 0, "Buffer too big! {}", buf.remaining());
-                    Ok(BigPingMsg { i, data, sum })
-                }
-                PONG_ID => Err(SerError::InvalidType(
-                    "Found BigPongMsg, but expected BigPingMsg.".into(),
-                )),
-                id => Err(SerError::InvalidType(format!(
-                    "Found unknown id {}, but expected BigPingMsg.",
-                    id
-                ))),
-            }
-        }
-    }
-
-    impl Deserialiser<BigPongMsg> for BigPingPongSer {
-        const SER_ID: SerId = Self::SID;
-
-        fn deserialise(buf: &mut dyn Buf) -> Result<BigPongMsg, SerError> {
-            if buf.remaining() < 18 {
-                return Err(SerError::InvalidData(format!(
-                    "Serialised typed has 18 bytes but only {} bytes remain in buffer.",
-                    buf.remaining()
-                )));
-            }
-            match buf.get_i8() {
-                PONG_ID => {
-                    let i = buf.get_u64();
-                    let data_len = buf.get_u64() as usize;
-                    let mut data = Vec::with_capacity(data_len as usize);
-                    for j in 0..data_len {
-                        data.insert(j, buf.get_u8());
-                    }
-                    let sum = buf.get_u64();
-                    assert_eq!(buf.remaining(), 0, "Buffer too big!");
-                    Ok(BigPongMsg { i, data, sum })
-                }
-                PING_ID => Err(SerError::InvalidType(
-                    "Found BigPingMsg, but expected BigPongMsg.".into(),
-                )),
-                id => Err(SerError::InvalidType(format!(
-                    "Found unknown id {}, but expected BigPingMsg.",
-                    id
-                ))),
-            }
-        }
-    }
-
-    #[derive(ComponentDefinition)]
-    struct BigPingerAct {
-        ctx: ComponentContext<BigPingerAct>,
-        target: ActorPath,
-        count: u64,
-        data_size: usize,
-        eager: bool,
-        buffer_config: BufferConfig,
-    }
-
-    #[allow(dead_code)]
-    impl BigPingerAct {
-        fn new_lazy(target: ActorPath, data_size: usize) -> BigPingerAct {
-            BigPingerAct {
-                ctx: ComponentContext::uninitialised(),
-                target,
-                count: 0,
-                data_size,
-                eager: false,
-                buffer_config: BufferConfig::default(),
-            }
-        }
-
-        fn new_eager(
-            target: ActorPath,
-            data_size: usize,
-            buffer_config: BufferConfig,
-        ) -> BigPingerAct {
-            BigPingerAct {
-                ctx: ComponentContext::uninitialised(),
-                target,
-                count: 0,
-                data_size,
-                eager: true,
-                buffer_config: buffer_config,
-            }
-        }
-    }
-
-    impl ComponentLifecycle for BigPingerAct {
-        fn on_start(&mut self) -> Handled {
-            let ping = BigPingMsg::new(0, self.data_size);
-            if self.eager {
-                self.ctx
-                    .init_buffers(Some(self.buffer_config.clone()), None);
-            }
-            debug!(self.ctx.log(), "Starting, sending first ping: {:?}", ping);
-            if self.eager {
-                self.target.tell_serialised(ping, self).expect("serialise");
-            } else {
-                self.target.tell(ping, self);
-            }
-            Handled::Ok
-        }
-    }
-
-    impl Actor for BigPingerAct {
-        type Message = Never;
-
-        fn receive_local(&mut self, _pong: Self::Message) -> Handled {
-            unimplemented!();
-        }
-
-        fn receive_network(&mut self, msg: NetMessage) -> Handled {
-            match msg.try_deserialise::<BigPongMsg, BigPingPongSer>() {
-                Ok(pong) => {
-                    debug!(self.ctx.log(), "Got msg {:?}", pong);
-                    pong.validate();
-                    assert_eq!(
-                        pong.data.len(),
-                        self.data_size,
-                        "Incorrect Pong data-length"
-                    );
-                    self.count += 1;
-                    if self.count < PING_COUNT {
-                        let ping = BigPingMsg::new(pong.i + 1, self.data_size);
-                        if self.eager {
-                            self.target.tell_serialised(ping, self).expect("serialise");
-                        } else {
-                            self.target.tell(ping, self);
-                        }
-                    }
-                }
-                Err(e) => error!(self.ctx.log(), "Error deserialising BigPongMsg: {:?}", e),
-            }
-            Handled::Ok
-        }
-    }
-
-    #[derive(ComponentDefinition)]
-    struct BigPongerAct {
-        ctx: ComponentContext<BigPongerAct>,
-        eager: bool,
-        buffer_config: BufferConfig,
-    }
-
-    #[allow(dead_code)]
-    impl BigPongerAct {
-        fn new_lazy() -> BigPongerAct {
-            BigPongerAct {
-                ctx: ComponentContext::uninitialised(),
-                eager: false,
-                buffer_config: BufferConfig::default(),
-            }
-        }
-
-        fn new_eager(buffer_config: BufferConfig) -> BigPongerAct {
-            BigPongerAct {
-                ctx: ComponentContext::uninitialised(),
-                eager: true,
-                buffer_config: buffer_config,
-            }
-        }
-    }
-
-    impl ComponentLifecycle for BigPongerAct {
-        fn on_start(&mut self) -> Handled {
-            if self.eager {
-                self.ctx
-                    .init_buffers(Some(self.buffer_config.clone()), None);
-            }
-            Handled::Ok
-        }
-    }
-
-    impl Actor for BigPongerAct {
-        type Message = Never;
-
-        fn receive_local(&mut self, _ping: Self::Message) -> Handled {
-            unimplemented!();
-        }
-
-        fn receive_network(&mut self, msg: NetMessage) -> Handled {
-            let sender = msg.sender;
-            match_deser! {msg.data; {
-                ping: BigPingMsg [BigPingPongSer] => {
-                    debug!(self.ctx.log(), "Got msg {:?} from {}", ping, sender);
-                    ping.validate();
-                    let pong = BigPongMsg::new(ping.i, ping.data.len());
-                    if self.eager {
-                        sender
-                            .tell_serialised(pong, self)
-                            .expect("BigPongMsg should serialise");
-                    } else {
-                        sender.tell(pong, self);
-                    }
-                },
-                !Err(e) => error!(self.ctx.log(), "Error deserialising BigPingMsg: {:?}", e),
-            }}
-            Handled::Ok
-        }
     }
 }
