+++ core/src/lib.rs
@@ -180,7 +180,7 @@ pub mod prelude {
             Require,
             RequireRef,
         },
-        net::buffers::ChunkLease,
+        net::buffers::{BufferConfig, ChunkLease, ChunkRef},
         ports::{Port, ProvidedPort, ProvidedRef, RequiredPort, RequiredRef},
         runtime::{KompactConfig, KompactSystem, SystemHandle},
         Never,
@@ -229,7 +229,7 @@ pub mod prelude {
 ///
 /// Import all with `use prelude_test::*;`.
 pub mod prelude_test {
-    pub use crate::serialisation::ser_test_helpers;
+    pub use crate::{net::net_test_helpers, serialisation::ser_test_helpers};
 }
 
 /// Helper structs and functions for doctests.
