+++ core/src/messaging/mod.rs
@@ -12,7 +12,7 @@ use uuid::Uuid;
 
 use crate::{
     net::{
-        buffers::{BufferChunk, ChunkLease},
+        buffers::{BufferChunk, ChunkLease, ChunkRef},
         frames::FRAME_HEAD_LEN,
     },
     serialisation::ser_helpers::deserialise_msg,
@@ -642,7 +642,9 @@ pub enum SerialisedFrame {
     /// Variant for Bytes allocated anywhere
     Bytes(Bytes),
     /// Variant for the Pooled buffers
-    Chunk(ChunkLease),
+    ChunkLease(ChunkLease),
+    /// Variant for the Pooled buffers
+    ChunkRef(ChunkRef),
 }
 
 impl SerialisedFrame {
@@ -654,7 +656,8 @@ impl SerialisedFrame {
     /// Returns the number of bytes in this frame
     pub fn len(&self) -> usize {
         match self {
-            SerialisedFrame::Chunk(chunk) => chunk.remaining(),
+            SerialisedFrame::ChunkLease(chunk) => chunk.remaining(),
+            SerialisedFrame::ChunkRef(chunk) => chunk.remaining(),
             SerialisedFrame::Bytes(bytes) => bytes.remaining(),
         }
     }
@@ -662,7 +665,8 @@ impl SerialisedFrame {
     /// Returns the data in this frame as a slice, in case of chaining only the front is returned!
     pub fn bytes(&self) -> &[u8] {
         match self {
-            SerialisedFrame::Chunk(chunk) => chunk.bytes(),
+            SerialisedFrame::ChunkLease(chunk) => chunk.bytes(),
+            SerialisedFrame::ChunkRef(chunk) => chunk.bytes(),
             SerialisedFrame::Bytes(bytes) => bytes.bytes(),
         }
     }
@@ -672,7 +676,10 @@ impl SerialisedFrame {
     pub fn make_contiguous(&mut self) {
         if self.len() > self.bytes().len() {
             match self {
-                SerialisedFrame::Chunk(chunk) => {
+                SerialisedFrame::ChunkLease(chunk) => {
+                    *self = SerialisedFrame::Bytes(chunk.to_bytes());
+                }
+                SerialisedFrame::ChunkRef(chunk) => {
                     *self = SerialisedFrame::Bytes(chunk.to_bytes());
                 }
                 _ => {
@@ -689,19 +696,13 @@ impl SerialisedFrame {
 pub enum DispatchData {
     /// Lazily serialised variant – must still be serialised by the dispatcher or networking system
     Lazy(Box<dyn Serialisable>),
-    /// Should be serialised and [framed](crate::net::frames).
-    Serialised((ChunkLease, SerId)),
+    /// Should be serialised and [framed](crate::net::frames::Frame).
+    SerialisedLease(ChunkLease),
+    /// Should be serialised and [framed](crate::net::frames::Frame).
+    SerialisedRef(ChunkRef),
 }
 
 impl DispatchData {
-    /// The serialisation id associated with the data
-    pub fn ser_id(&self) -> SerId {
-        match self {
-            DispatchData::Lazy(ser) => ser.ser_id(),
-            DispatchData::Serialised((_chunk, ser_id)) => *ser_id,
-        }
-    }
-
     /// Try to extract a network message from this data for local delivery
     ///
     /// This can fail, if the data can't be moved onto the heap, and serialisation
@@ -712,7 +713,13 @@ impl DispatchData {
                 let ser_id = ser.ser_id();
                 Ok(NetMessage::with_box(ser_id, src, dst, ser))
             }
-            DispatchData::Serialised((mut chunk, _ser_id)) => {
+            DispatchData::SerialisedLease(mut chunk) => {
+                // The chunk contains the full frame, deserialize_msg does not deserialize FrameHead so we advance the read_pointer first
+                chunk.advance(FRAME_HEAD_LEN as usize);
+                //println!("to_local (from: {:?}; to: {:?})", src, dst);
+                Ok(deserialise_msg(chunk).expect("s11n errors"))
+            }
+            DispatchData::SerialisedRef(mut chunk) => {
                 // The chunk contains the full frame, deserialize_msg does not deserialize FrameHead so we advance the read_pointer first
                 chunk.advance(FRAME_HEAD_LEN as usize);
                 //println!("to_local (from: {:?}; to: {:?})", src, dst);
@@ -729,10 +736,11 @@ impl DispatchData {
         buf: &mut BufferEncoder,
     ) -> Result<SerialisedFrame, SerError> {
         match self {
-            DispatchData::Lazy(ser) => Ok(SerialisedFrame::Chunk(
+            DispatchData::Lazy(ser) => Ok(SerialisedFrame::ChunkLease(
                 crate::serialisation::ser_helpers::serialise_msg(&src, &dst, ser.deref(), buf)?,
             )),
-            DispatchData::Serialised((chunk, _ser_id)) => Ok(SerialisedFrame::Chunk(chunk)),
+            DispatchData::SerialisedLease(chunk) => Ok(SerialisedFrame::ChunkLease(chunk)),
+            DispatchData::SerialisedRef(chunk) => Ok(SerialisedFrame::ChunkRef(chunk)),
         }
     }
 }
